kp: Do not mark kp_notifier_callback() as an inline function
An inline function is one for which the compiler copies the code from the
function definition directly into the code of the calling function rather
than creating a separate set of instructions in memory. This is mostly done
for small computations, means a function having less codes so that we can
ignore its overhead and perform faster execution.

Here, kp_notifier_callback() is by no means a small function definition. The
compiler will disregard the inline function hint and optimize it. Therefore,
marking the function inline is pointless and only becomes a burden for the
compiler to recognize the function's body types and thereafter optimize it.
Remove the inline specifier.

This reverts commit f80bb0450c69edad5d982a7de8a6c73bbd67b23b.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>